+++ lib/src/widgets/scaffold.dart
@@ -86,7 +86,7 @@ class WaveScaffold extends StatelessWidget {
       extendBody: extendBody,
       extendBodyBehindAppBar: extendBodyBehindAppBar,
       appBar: appBar,
-      body: body,
+      body: body == null ? null : SafeArea(child: body!),
       floatingActionButton: floatingActionButton,
       floatingActionButtonLocation: floatingActionButtonLocation,
       floatingActionButtonAnimator: floatingActionButtonAnimator,
@@ -127,59 +127,19 @@ class WaveScaffold extends StatelessWidget {
           floatingActionButtonAnimator,
         ),
       )
-      ..add(
-        DiagnosticsProperty<bool>(
-          'extendBodyBehindAppBar',
-          extendBodyBehindAppBar,
-        ),
-      )
+      ..add(DiagnosticsProperty<bool>('extendBodyBehindAppBar', extendBodyBehindAppBar))
       ..add(DiagnosticsProperty<bool>('extendBody', extendBody))
-      ..add(
-        DiagnosticsProperty<AlignmentDirectional>(
-          'persistentFooterAlignment',
-          persistentFooterAlignment,
-        ),
-      )
-      ..add(
-        ObjectFlagProperty<DrawerCallback?>.has(
-          'onDrawerChanged',
-          onDrawerChanged,
-        ),
-      )
-      ..add(
-        ObjectFlagProperty<DrawerCallback?>.has(
-          'onEndDrawerChanged',
-          onEndDrawerChanged,
-        ),
-      )
+      ..add(DiagnosticsProperty<AlignmentDirectional>('persistentFooterAlignment', persistentFooterAlignment))
+      ..add(ObjectFlagProperty<DrawerCallback?>.has('onDrawerChanged', onDrawerChanged))
+      ..add(ObjectFlagProperty<DrawerCallback?>.has('onEndDrawerChanged', onEndDrawerChanged))
       ..add(ColorProperty('drawerScrimColor', drawerScrimColor))
       ..add(ColorProperty('backgroundColor', backgroundColor))
-      ..add(
-        DiagnosticsProperty<bool?>(
-          'resizeToAvoidBottomInset',
-          resizeToAvoidBottomInset,
-        ),
-      )
+      ..add(DiagnosticsProperty<bool?>('resizeToAvoidBottomInset', resizeToAvoidBottomInset))
       ..add(DiagnosticsProperty<bool>('primary', primary))
-      ..add(
-        EnumProperty<DragStartBehavior>(
-          'drawerDragStartBehavior',
-          drawerDragStartBehavior,
-        ),
-      )
+      ..add(EnumProperty<DragStartBehavior>('drawerDragStartBehavior', drawerDragStartBehavior))
       ..add(DoubleProperty('drawerEdgeDragWidth', drawerEdgeDragWidth))
-      ..add(
-        DiagnosticsProperty<bool>(
-          'drawerEnableOpenDragGesture',
-          drawerEnableOpenDragGesture,
-        ),
-      )
-      ..add(
-        DiagnosticsProperty<bool>(
-          'endDrawerEnableOpenDragGesture',
-          endDrawerEnableOpenDragGesture,
-        ),
-      )
+      ..add(DiagnosticsProperty<bool>('drawerEnableOpenDragGesture', drawerEnableOpenDragGesture))
+      ..add(DiagnosticsProperty<bool>('endDrawerEnableOpenDragGesture', endDrawerEnableOpenDragGesture))
       ..add(StringProperty('restorationId', restorationId));
   }
 }