+++ lib/waveui.dart
@@ -4,6 +4,7 @@ library waveui;
 export 'assets.dart';
 export 'foundation.dart';
 export 'theme.dart';
+export 'utils.dart';
 
 export 'widgets/app_bar.dart';
 export 'widgets/app.dart';
@@ -16,3 +17,4 @@ export 'widgets/pull_to_refresh.dart';
 export 'widgets/progress.dart';
 export 'widgets/scaffold.dart';
 export 'widgets/form.dart';
+export 'widgets/file_picker.dart';