+++ lib/widgets/file_picker.dart
@@ -1,0 +1,2 @@
+export '../src/widgets/file_picker/file_picker.dart';
+export '../src/widgets/file_picker/file_picker_item.dart';