+++ CHANGELOG.md
@@ -27,7 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - badge
 - navigation_bar
 - text_form_field
-
+- file_picker and file_picker_item
+- file_utils
 ### Changed
 
 ### Deprecated